fix bug in create method for attachments controller
I believe the issue is that when I was creating a "new" item for the authorization to validate against, that new item was added to the relation and when the attach function was called, the whole relation was saved, not just the one attachment, thus attempting to save an empty attachment, creating an invalid error.  This must be new with the more recent versions of this gem.